♻️ refactor: split challenge page components #24
